Expose PSPModule (so the debugger can access it later)
